requirements.txt
# gcloud run deploy smart-doc-searcher-api --source . --region us-central1 --allow-unauthenticated --timeout 300 --cpu-throttling --max-instances 1 --concurrency 1
     
Flask==3.0.3
gunicorn==22.0.0
google-genai==0.1.0
google-cloud-storage==2.16.0
pdfminer.six
python-docx
